requirements.txt
python-dotenv==0.19.1

Flask==2.3.0
Werkzeug==2.3.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1

SQLAlchemy==1.4.37
alembic==1.8.0

PyJWT==1.7.1

pytest==7.1.2
pytest-mock==2.0.0
pytest-cov==2.12.1
psycopg2==2.9.5